UPSTREAM: Revert "lint/clang-format: set to 96 chars per line"
This reverts commit 626ba097a2cd1c87800a2154420829b09803467e.

This change was submitted under the incorrect assumption that there was
agreement on a coding style change. There wasn't, so while the issue is
under discussion we should revert to the previous status quo.

Making clang-format honor the line length is a separate issue from
changing the line length, and can be reuploaded as a separate CL.

BUG=none
BRANCH=none
TEST=none

Change-Id: I19d7f79e8cfefe50d1d3ccb25f2daf02a02df100
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 0e3f7d47804d39912faec5a56bc9cddf91ea354b
Original-Change-Id: I433c82c95a897b3113cace3668cc8ce0f1ab75bf
Original-Signed-off-by: Julius Werner <jwerner@chromium.org>
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/31916
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Reviewed-on: https://chromium-review.googlesource.com/1524800
Commit-Ready: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>