UPSTREAM: soc/intel/common: Update ESPI disable option
Update the Kconfig option for disabling ESPI SMI source
to disable it entirely, not just when ACPI mode is disabled.

For the situations where this is needed (just the sarien
board) it is better to completely stop the EC from sending
any SMI events as no actions are taken.

BUG=none
BRANCH=none
TEST=none

Change-Id: I1cd70c5184b32e4e4f39eb049a4e5dfbf3fcca85
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: af39c82a36119158ffcee955cf83d880c9560e5f
Original-Change-Id: Id94481bb2f0cfc948f350be45d360bfe40ddf018
Original-Signed-off-by: Duncan Laurie <dlaurie@google.com>
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/31918
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Furquan Shaikh <furquan@google.com>
Reviewed-on: https://chromium-review.googlesource.com/1528218
Commit-Ready: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>