UPSTREAM: gale: add dev switch back as physical presence GPIO
gale has a button which is essentially used as a
"physical presence" button.  Its only use is to emulate
^D or ^U on boot when the button is pressed.
(See depthcharge src/board/gale/board.c)

Previously (and currently in CrOS firmware branch) this
GPIO was defined as the physical developer switch,
and read as such in depthcharge.  It was removed in
cleanup patch CB:18980.

Add the GPIO back as physical presence ("presence"),
which will be read by depthcharge in CL:1532492.

BUG=b:124141368, b:124192753, chromium:942901
TEST=util/lint/checkpatch.pl -g origin/master..HEAD
TEST=util/abuild/abuild -B -e -y -c 50 -p none -x
TEST=make clean && make test-abuild
BRANCH=none

Change-Id: Idbc92c9fb6f1dcbf9e050937696945707f3dc6c6
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 99625b0cd83871b30aca6ad7bb182326f8467673
Original-Change-Id: Ic144f839b7f9933d573db8f84c4bf5905eea96f6
Original-Signed-off-by: Joel Kitching <kitching@google.com>
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/31993
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Julius Werner <jwerner@chromium.org>
Original-Reviewed-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1539547
Commit-Ready: Furquan Shaikh <furquan@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>