UPSTREAM: mainboard/google/sarien: skip tpm check when !verstage
The TPM driver isn't loaded in other stages but verstage so when we try
to communicate with the TPM it fails. We don't need to communicate with
it anyway since the TPM won't continue to tell us that recovery was
requested, only the first query responds with the recovery request.

BRANCH=none
BUG=b:129150074,b:123360379
TEST=1)boot arcada without recovery and notice that the "tpm transaction
failed" log lines are no longer present. 2) boot into recovery using the
ESC refresh power key combination and verify that the recovery reason
was "recovery button pressed"

Change-Id: I2db902d46eba72cf4799dee3620193c30b5af1dc
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 837f6552917852d618b5dbce960e95ede3c3fe26
Original-Change-Id: I13284483d069ed50b0d16b36d0120d006485f7f4
Original-Signed-off-by: Jett Rink <jettrink@chromium.org>
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/32030
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Keith Short <keithshort@chromium.org>
Original-Reviewed-by: Duncan Laurie <dlaurie@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1539546
Commit-Ready: Furquan Shaikh <furquan@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>