UPSTREAM: nb/intel/pineview: Correct lsbpos(0) and msbpos(0)
lsbpos and msbpos have incorrect behaviour when given 0.
lsbpos(0) returns 8, and msbpos(0) hangs. The latter is
because the check i >= 0 is always true for an unsigned
integer, causing it to loop indefinitely (this was flagged
by Coverity).

0 doesn't have a lsb or msb position, so we change both
functions to return -1 in this case to indicate an error.
The code already guards against calling these functions
with 0, but we make this more explicit to prevent errors
in the future.

BUG=none
BRANCH=none
TEST=none

Change-Id: I61f03e224603ccde1f6677fd29e87353372c389b
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: b70c77691ba9a26989fc8922a2e1807f6f8bdd09
Original-Found-by: Coverity Scan, CID 1347356, 1347386
Original-Signed-off-by: Jacob Garber <jgarber1@ualberta.ca>
Original-Change-Id: Ic5be50846cc545dcd48593e5ed3fd6068a6104cb
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/32054
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Paul Menzel <paulepanter@users.sourceforge.net>
Original-Reviewed-by: Patrick Georgi <pgeorgi@google.com>
Reviewed-on: https://chromium-review.googlesource.com/1558128
Commit-Ready: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>