UPSTREAM: mediatek/mt8183: Calibrate vsim2 to 2.7 V
The default voltage of vsim2 is set to 2.76V for sim card usage.
In general, 2.76V of vsim2 is composed of 2.7V main voltage and 0.06V calibration voltage.
However, vsim2 is used for the tx_ovdd power of display port IT6505 on the kukui board design which needs 2.7V.
So we set it to 2.7V with modifying calibration value.

BUG=b:126139364
BRANCH=none
TEST=measure vsim2 voltage with multimeter

Change-Id: I4c3a91d8bfebb705df00332335ee07066c1c8591
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: c10af299aeed2a7d530fbabbaa55947d64712333
Original-Change-Id: I4dffdde89cbde91286d92e6c2b445f0b3d0ad2fe
Original-Signed-off-by: Hsin-Hsiung Wang <hsin-hsiung.wang@mediatek.com>
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/32057
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Julius Werner <jwerner@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/1672267
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>