UPSTREAM: soc/intel/cannonlake: Update CPU Ratio base on MSR
The following is the FSP logic: as long as the Cpu Ratio input in
coreboot is different with CpuStrapSet, system will force to follow
input from coreboot. But CpuStrapsetting is floating, it will be 0
from the first cold boot before memory training and set to 0x1c (or
max CPU ratio for the installed CPU) after first memory training.

The previous fix was attempting to ensure settings were cleared
when FSP was called in recovery mode, but only when coming from S5
which caused issues if recovery mode is requested by the OS and
is only followed by a warm reset.

BUG=b:129412691
TEST=Boot up sarien platform and force recovery, check there's no reset
in the path of recovery.

Change-Id: Iff098f78876e9fae50adf3d02467131715db98bf
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 68890b9d59d0f87533d3eacb6de1d552dcb72679
Original-Signed-off-by: Lijian Zhao <lijian.zhao@intel.com>
Original-Change-Id: I959188be46343bc6f2cb3cc149097b4d449802aa
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/32089
Original-Reviewed-by: Furquan Shaikh <furquan@google.com>
Original-Reviewed-by: Duncan Laurie <dlaurie@chromium.org>
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/1545769
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Duncan Laurie <dlaurie@google.com>
Commit-Queue: Simon Glass <sjg@chromium.org>