UPSTREAM: riscv: add support for OpenSBI
Call OpenSBI in M-Mode and use it to set up SBI and to lockdown the
platform. It will also jump to the specified payload when done.
This behaviour is similar to BL31 on aarch31.

The payload is 41KiB in size on qemu.

Tested on qemu-riscv:
Required to boot a kernel as OpenSBI's instruction emulation feature
is required on that virtual machine.

Tested on SiFive/unleashed:
The earlycon is working. No console after regular serial driver
should take over, which might be related to kernel config.

BUG=none
BRANCH=none
TEST=none

Change-Id: Ifba593ebefa0e83f979acb77cb79dfbd89080196
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: a6f9eab44ab0590ca7da33da0b042a8fce8da0f1
Original-Change-Id: I2a178595bd2aa2e1f114cbc69e8eadd46955b54d
Original-Signed-off-by: Xiang Wang <merle@hardenedlinux.org>
Original-Signed-off-by: Patrick Rudolph <patrick.rudolph@9elements.com>
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/32394
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Philipp Deppenwiese <zaolin.daisuki@gmail.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/1734793
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>