UPSTREAM: mb/google/hatch: Re-configure GPP_A12 as GPO before entering sleep
GPP_A12 has a Native3 (SX_EXIT_HOLDOFF#) mode, which allows to delay
resuming to S0. If this pad is not locked and platform was not initially
designed for this functionality, malware could reconfigure this pads
setting under OS (switch to Native3), which would make platform not able
to resume until G3 is applied. To prevent misuse of this pad,
re-configure this pad before entering S3 and S5 to guarantee that the
pad configuration is correct.

BUG=b:128686027

Change-Id: If0a8e5fb0df00931cdcd083cd03a2f15832556cb
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: c41b0e8ea3b2714bf6d6d88a6b66c333c6919f07
Original-Change-Id: I1e7979baa491acf2c56d223afb4618f0f6429e37
Original-Signed-off-by: Krishna Prasad Bhat <krishna.p.bhat.d@intel.com>
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/32111
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Furquan Shaikh <furquan@google.com>
Reviewed-on: https://chromium-review.googlesource.com/1547566
Commit-Ready: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>