UPSTREAM: arch/x86/smbios: Add type 7
The SMBIOS spec requires type 7 to be present.

Add the type 7 fields and enums for SMBIOS 3.1+ and fill it with the
"Deterministic Cache Parameters" as available on Intel and AMD.

As CPUID only provides partial information on caches, some fields are set to
unknown.
The following fields are supported:
* Cache Level
* Cache Size
* Cache Type
* Cache Ways of Associativity

Tested on Intel Sandy Bridge (Lenovo T520).
All 4 caches are displayed in dmidecode and show the correct information.

BUG=none
BRANCH=none
TEST=none

Change-Id: I36f615b81403665f4bef1e337d872bbc99588ca2
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: fc5b80943b849ae3f949c8647aca5bb91872e4a7
Original-Change-Id: I80ed25b8f2c7b425136b2f0c755324a8f5d1636d
Original-Signed-off-by: Patrick Rudolph <patrick.rudolph@9elements.com>
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/32131
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Richard Spiegel <richard.spiegel@silverbackltd.com>
Reviewed-on: https://chromium-review.googlesource.com/1560240
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>