UPSTREAM: soc/skl: Update SkipExtGfxScan in UPD from devtree
The SkipExtGfxScan option is defined in the device tree, but doesn`t
update the value in the UPD. It uses the default value - 0. This
means that the FSP will scan all external graphics devices, in spite
of the configuration in devicetree.cb for a specific board.

Patch updates SkipExtGfxScan options in UPD from devicetree.cb.
This change affects all boards with skl/kbl processor.

BUG=none
BRANCH=none
TEST=none

Change-Id: Iccb2af3c6f28579daf78d5247b3fd745e36cde8e
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: a12e9b0666e1a22958ea1fae87199a8b82266ee7
Original-Change-Id: Ie88a41bdf31f7c3e88df6c70c82a1cbf866372c4
Original-Signed-off-by: Maxim Polyakov <max.senia.poliak@gmail.com>
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/32170
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Nico Huber <nico.h@gmx.de>
Reviewed-on: https://chromium-review.googlesource.com/1558125
Commit-Ready: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>