UPSTREAM: google/kukui: Include LCD module identifier (LCM ID) into SKU ID
Kukui is using MIPI display panel and needs some identifier to tell
payloads which LCD module is installed, and to select right kernel
device tree. Following Scarlet, the decision is to embed LCD module ID
as part of SKU ID.

The LCM ID is using a different voltage mapping table from the rest.
Considering the complexity in computation of SKU ID, it is better to
move the cache logic from get_index to caller.

Also revise the mapping table since ADC on 8183 only supports 12
levels.

BUG=b:129299873
TEST=make -j; boots on Kukui Rev2 unit.

Change-Id: I3499869f058f89f02fdfb09fd492e81d885c6515
Signed-off-by: Martin Roth <martinroth@chromium.org>
Original-Commit-Id: 59a407349b6ed15fd83023b96ee559b7771c52f8
Original-Change-Id: Ib0c00bc8ce3c71c445c5c4561403ce8ef4dd5844
Original-Signed-off-by: Hung-Te Lin <hungte@chromium.org>
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/32263
Original-Reviewed-by: Julius Werner <jwerner@chromium.org>
Original-Reviewed-by: Paul Menzel <paulepanter@users.sourceforge.net>
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Reviewed-on: https://chromium-review.googlesource.com/1569033
Reviewed-by: Hung-Te Lin <hungte@chromium.org>