UPSTREAM: mb/siemens/mc_apl1: Enable HDA in devicetree for all mainboard variants
With commit
'4074ce0cc7 (intel/apollolake: Add HDA to disable_dev function)'
FSP is now requested to switch off HDA PCI device if it is disabled in
devicetree. Doing so results in a warm restart. Normally this event
will be stored in CMOS RAM (if the descriptor is configured to do so)
and therefore no further resets are requested by FSP on the next boots
as long as CMOS RAM is kept alive.

The Siemens mainboards based on Apollo Lake do not have a CMOS battery
and therefore the CMOS is not backed up. This leads to reset requests
from FSP after PCI enumeration on every boot. To avoid this reset enable
HDA in devicetree for these mainboards. Though we do not have any usage
of HDA it should not be an issue that the HDA device is now enabled. The
benefit is though that no reset is requested anymore by FSP.

BUG=none
BRANCH=none
TEST=none

Change-Id: I8068e4a4c238c3b4064ca650bbf29dffb6523272
Signed-off-by: Martin Roth <martinroth@chromium.org>
Original-Commit-Id: a4e5236e8945d0acab13e9d6c69daf1efa7ea726
Original-Change-Id: I637c7c01d73350700c6066fee74fecbb5b93b221
Original-Signed-off-by: Werner Zeh <werner.zeh@siemens.com>
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/32295
Original-Reviewed-by: Kysti Mlkki <kyosti.malkki@gmail.com>
Original-Reviewed-by: Uwe Poeche <uwe.poeche@siemens.com>
Original-Reviewed-by: Mario Scheithauer <mario.scheithauer@siemens.com>
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Reviewed-on: https://chromium-review.googlesource.com/1567860