UPSTREAM: libpayload: classify all keyboards
Depthcharge uses the keyboard type to help determine whether
it can trust the keyboard for security-sensitive confirmations.
Currently it trusts anything except usb, but now there's a need
to distrust ec-based ps/2 keyboards that are associated with untrusted
ECs.  To help facilitate this, coreboot needs to report more
details about non-usb keyboards, so this change replaces the current
instances of unknown with enum values that distinguish uart and gpio
from ec-based keyboards.

BUG=b:129471321
BRANCH=None
TEST=Local compile and flash to systems with trusted and non-trusted
ECs.  Confirmed that security confirmation can't be performed via
keyboard on a system with an untrusted EC but can still be performed
on a system with a trusted EC.

Change-Id: I7243fbbd331bca5e80f67c9c86725d3f9aefdd83
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: a20e59da157231f0a0e41774af80f6bccde1b280
Original-Change-Id: Iee6295dafadf7cb3da98b62f43b0e184b2b69b1e
Original-Signed-off-by: Matt Delco <delco@chromium.org>
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/32717
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Julius Werner <jwerner@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1609441
Commit-Ready: Matt Delco <delco@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>