UPSTREAM: soc/amd/picasso: Create picasso as a copy of stoneyridge
So that everyone can see what's being updated from stoney, we're
starting with a direct copy of the stoney directory. There are
arguments both for and against doing it this way, but I believe
This the most transparent way.  We've moved much of the duplicated
stoney code into the soc/amd/common directory and will continue
that work as it becomes obvious that we have unchanged code between
the SOCs.

Makefile.inc has been renamed as makefile.inc so that it won't
build in jenkins until the directory is  updated.

Other than that change, this is an exact copy of the stoneyridge
SOC directory which will be updated in the follow-on commits in
the patch train.

TEST=None
BUG=b:130804851

Signed-off-by: Martin Roth <gaumless@gmail.com>
Original-Commit-Id: 5c354b9979c7e7ad9af668bad0e1b6a5c3003d26
Original-Signed-off-by: Martin Roth <martinroth@chromium.org>
Original-Change-Id: I6809bd1eea304f76dd9000c079b3ed09f94dbd3b
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/32407
Original-Reviewed-by: Richard Spiegel <richard.spiegel@silverbackltd.com>
Original-Reviewed-by: Angel Pons <th3fanbus@gmail.com>
Original-Reviewed-by: Edward O'Callaghan <quasisec@chromium.org>
Original-Reviewed-by: HAOUAS Elyes <ehaouas@noos.fr>
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Change-Id: Ica6e9ed918c444392755040931eebc653252f6d7
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/1686883
Reviewed-by: Martin Roth <martinroth@chromium.org>
Tested-by: Martin Roth <martinroth@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>