UPSTREAM: arcada: add internal pull to ISH UART RX
We do not want the RX signal to be floating on the board as that could
cause the ISH to remain in a higher power state (because there is logic
to keep the ISH in an higher power state when there is an active UART).

Add an internal 20K pull up on the RX line. In normal configuration this
will burn an additional 544uW.

BRANCH=R75
BUG=b:131241969
TEST=verify that ISH console still works with rework

Change-Id: If7940dd4ae9b90cd9e96cafff7b016af28868aac
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: c426be6ae29ad345d5f6a42726859fd8457f9814
Original-Change-Id: Ifc9621bcafe4c86edfa9cd6d58b307254d3a81ca
Original-Signed-off-by: Jett Rink <jettrink@chromium.org>
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/32471
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Furquan Shaikh <furquan@google.com>
Original-Reviewed-by: Duncan Laurie <dlaurie@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1587978
Commit-Ready: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>